index.js: a minor correction
Number("5") = 5

The Number() attempts to convert a string into a number.
If it succeeds, it will return the numeric value.
If it fails, it will return NaN, which you can then test for
using the function isNaN("abcd").

When you use a prompt, it will return a string.
When you check for equality, it's a good idea to make sure your
types match instead of letting javascript try to make that
decision for you.

Javascript and numbers can be very fickle, especially
when you start playing around with numbers and math!